+++ src/librados/librados.cc
@@ -1309,7 +1309,7 @@ int librados::IoCtx::aio_operate(const std::string& oid, AioCompletion *c,
 				  translate_flags(flags));
 }
 
-int librados::IoCtx::aio_operate(const std::string& oid, AioCompletion *c,
+uint64_t librados::IoCtx::aio_operate(const std::string& oid, AioCompletion *c,
 				 librados::ObjectWriteOperation *o,
 				 snap_t snap_seq, std::vector<snap_t>& snaps)
 {
